Trait backup_cli::storage::BackupStorage

pub trait BackupStorage {
#[must_use]
    fn create_backup<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        name: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<BackupHandle>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn create_for_write<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        backup_handle: &'life1 BackupHandleRef, 
        name: &'life2 str
    ) -> Pin<Box<dyn Future<Output = Result<(FileHandle, Box<dyn AsyncWrite + Send + Unpin>)>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn open_for_read<'life0, 'async_trait>(
        file_handle: &'life0 FileHandleRef
    ) -> Pin<Box<dyn Future<Output = Result<Box<dyn AsyncRead + Send + Unpin>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait;
}

Required methods

#[must_use] fn create_backup<'life0, 'life1, 'async_trait>(
&'life0 self,
name: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<BackupHandle>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Hint that a bunch of files are gonna be created related to a backup identified by name, which is unique to the content of the backup, i.e. it won't be the same name unless you are backing up exactly the same thing. Storage can choose to take actions like create a dedicated folder or do nothing. Returns a string to identify this operation in potential succeeding file creation requests.

#[must_use] fn create_for_write<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
backup_handle: &'life1 BackupHandleRef,
name: &'life2 str
) -> Pin<Box<dyn Future<Output = Result<(FileHandle, Box<dyn AsyncWrite + Send + Unpin>)>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Ask to create a file for write, backup_handle was returned by create_backup to identify the current backup.

#[must_use] fn open_for_read<'life0, 'async_trait>(
file_handle: &'life0 FileHandleRef
) -> Pin<Box<dyn Future<Output = Result<Box<dyn AsyncRead + Send + Unpin>>> + Send + 'async_trait>> where
'life0: 'async_trait,

Open file for reading.

Loading content...